Code refactoring + README
