--- walkthroughs/cdk-examples/howto-alb-cdk/colorapp/requirements.txt
@@ -1,3 +1,0 @@
-flask
-aws-xray-sdk
-gunicorn==19.9.0
