+++ walkthroughs/cdk-examples/howto-alb/cdk.json
