+++ walkthroughs/cdk-examples/howto-alb/colorapp/Dockerfile
