+++ walkthroughs/cdk-examples/howto-alb/colorapp/config.py
