+++ walkthroughs/cdk-examples/howto-alb/feapp/Dockerfile
