+++ walkthroughs/cdk-examples/howto-alb/lib/stacks/ecs-services.ts
@@ -1,4 +1,4 @@
-import { StackProps, Stack } from "aws-cdk-lib";
+import { StackProps, Stack, CfnOutput } from "aws-cdk-lib";
 import { BackendServiceV2Construct } from "../constructs/backend-service-v2";
 import { BackendServiceV1Construct } from "../constructs/backend-service-v1";
 import { FrontEndServiceConstruct } from "../constructs/frontend-service";
@@ -12,11 +12,14 @@ export class ECSServicesStack extends Stack {
     const backendV2 = new BackendServiceV2Construct(ms, "BackendServiceV2Construct");
     const frontend = new FrontEndServiceConstruct(ms, "FrontEndServiceConstruct");
 
-    backendV1.service.node.addDependency(backendV2.service);
-    frontend.service.node.addDependency(backendV2.service);
-    frontend.service.node.addDependency(backendV1.service);
+    // backendV1.service.node.addDependency(backendV2.service);
+    // frontend.service.node.addDependency(backendV2.service);
+    // frontend.service.node.addDependency(backendV1.service);
 
     frontend.taskDefinition.node.addDependency(ms.frontendVirtualNode);
     frontend.taskDefinition.node.addDependency(ms.backendVirtualService);
+
+    new CfnOutput(this, 'PublicEndpoint', { value: ms.sd.frontendLoadBalancer.loadBalancerDnsName, 
+    description: 'Public endpoint to query the frontend load balancer', exportName: 'PublicEndpoint' });
   }
 }
