+++ walkthroughs/cdk-examples/howto-alb/tsconfig.json
